+++ src/components/AccountDrawer/SettingsMenu.tsx
@@ -4,7 +4,7 @@ import Row from 'components/Row'
 import { LOCALE_LABEL } from 'constants/locales'
 import { useCurrencyConversionFlagEnabled } from 'featureFlags/flags/currencyConversion'
 import { useActiveLocale } from 'hooks/useActiveLocale'
-import { useActiveLocalCurrency } from 'hooks/useLocalActiveCurrency'
+import { useActiveLocalCurrency } from 'hooks/useActiveLocalCurrency'
 import { ReactNode } from 'react'
 import { ChevronRight } from 'react-feather'
 import styled from 'styled-components'
